Add files via upload
Everything is now under database configuration rather than from temp. directory of the previous gen, and some tweaks are needed to fully integrate it but up to 90% is already done on the DB struc. just needs some adjustments for other advanced features, such as advanced analytics and other things which are found outside of the app.py main function and UI interface